Bifurcate non-ROS and ROS templates
I'm now going to be trying to make the ROS templates without a ROS setup, so
I'm pressing forwards with making them happen.

The previous non-ROS template is now 'animate', with the idea being that
as a side-project I might implement a test animator using it.  But this isn't
a priority.